% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datExAltServer.R
\name{datExAltServer}
\alias{datExAltServer}
\title{Display score at an age for the data exploration page}
\usage{
datExAltServer(id, modelDataEdit, modelFit, modelType, traj)
}
\description{
Display score at an age for the data exploration page
}
\keyword{internal}
