% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datExAltUI.R
\name{datExAltUI}
\alias{datExAltUI}
\title{Display score at an age for the data exploration page}
\usage{
datExAltUI(id, label = "Model")
}
\description{
Display score at an age for the data exploration page
}
\keyword{internal}
