% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datExAUCServer.R
\name{datExAUCServer}
\alias{datExAUCServer}
\title{AUC for the data exploration page}
\usage{
datExAUCServer(id, modelDataEdit, modelFit, modelType, traj, button)
}
\description{
AUC for the data exploration page
}
\keyword{internal}
