% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadExploreServer.R
\name{downloadExploreServer}
\alias{downloadExploreServer}
\title{Server to download results from explore data page}
\usage{
downloadExploreServer(
  id,
  descTable,
  warningMsg,
  formCodeRender,
  statement,
  fixedTab,
  interpretation,
  interpretationRand,
  randomTab,
  N,
  mainPlot,
  phenotype,
  modelType,
  datExAltTable,
  datExAltPlot,
  AUC,
  plotAUC,
  tableAUC
)
}
\description{
Server to download results from explore data page
}
\keyword{internal}
