% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadExploreUI.R
\name{downloadExploreUI}
\alias{downloadExploreUI}
\title{UI to download results from explore data page}
\usage{
downloadExploreUI(id, label = "data")
}
\description{
UI to download results from explore data page
}
\keyword{internal}
