% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importantAgeServer.R
\name{importantAgeServer}
\alias{importantAgeServer}
\title{split model by variable}
\usage{
importantAgeServer(id, modelDataEdit, modelType, modelFit, age)
}
\description{
split model by variable
}
\keyword{internal}
