% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchTIDAL.R
\name{launchTIDAL}
\alias{launchTIDAL}
\title{Run Shiny app from local R environment}
\usage{
launchTIDAL(display = "default")
}
\arguments{
\item{display}{Character vector. Can be either "default" or "browser". Default is "default".
"default" opens the Shiny app in a pop-out window in RStudio, while "browser" opens it in a web browser.}
}
\description{
Opens an interactive Shiny GUI.
}
\examples{
\dontrun{# Launch Shiny app with default display
launchTIDAL()

# Launch Shiny app with browser display
launchTIDAL(display = "browser")
}

You can also manually open a browser and copy and paste the URL after "Listening on" into the address bar.
Eg. for "Listening on http://127.0.0.1:4484" copy and paste "http://127.0.0.1:4484" into your browser, eg. in Chrome or Firefox.
Even though you are using a browser the app is still running locally on your computer, the same as if the pop-out window was launched from RStudio.

}
