% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCondServer.R
\name{modelCondServer}
\alias{modelCondServer}
\title{split model by variable}
\usage{
modelCondServer(
  id,
  modelData,
  formCodeCovars,
  dfPlot,
  traj,
  age,
  timePoint,
  modelType
)
}
\description{
split model by variable
}
\keyword{internal}
