% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlotServer.R
\name{modelPlotServer}
\alias{modelPlotServer}
\title{Model plot - data exploration page}
\usage{
modelPlotServer(
  id,
  modelData,
  modelFit,
  age,
  traj,
  timePoint,
  modelType,
  button
)
}
\description{
Model plot - data exploration page
}
\keyword{internal}
