% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelResultsServer.R
\name{modelResultsServer}
\alias{modelResultsServer}
\title{Run model - data exploration page}
\usage{
modelResultsServer(
  id,
  modelFit,
  warningMsg,
  modelData,
  modelType,
  randomFX,
  age,
  traj,
  covars
)
}
\description{
Run model - data exploration page
}
\keyword{internal}
