% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelRunServer.R
\name{modelRunServer}
\alias{modelRunServer}
\title{Run model - data exploration page}
\usage{
modelRunServer(
  id,
  covariateChoice,
  button,
  modelData,
  formCodeCovars,
  age,
  traj,
  timePoint
)
}
\description{
Run model - data exploration page
}
\keyword{internal}
