% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDataServer.R
\name{selectDataServer}
\alias{selectDataServer}
\title{Shiny module - Selecting and reading in the dataset}
\usage{
selectDataServer(id, dataFormatted)
}
\description{
Shiny module - Selecting and reading in the dataset
}
\keyword{internal}
