% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleTrajServer.R
\name{singleTrajServer}
\alias{singleTrajServer}
\title{Server for individual trajectories}
\usage{
singleTrajServer(id, subject, age, traj, modelData, modelFit, modelType, cov)
}
\description{
Server for individual trajectories
}
\keyword{internal}
