% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2longServer.R
\name{wide2longServer}
\alias{wide2longServer}
\title{This function takes a wide format dataset and converts it to a long format dataset.}
\arguments{
\item{id}{The id of the module server}
}
\value{
Returns a long format dataset
}
\description{
This function takes a wide format dataset and converts it to a long format dataset.
}
\keyword{internal}
