% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2longUI.R
\name{wide2longUI}
\alias{wide2longUI}
\title{This function takes a wide format dataset and converts it to a long format dataset.}
\usage{
wide2longUI(id, label = "wide2long")
}
\description{
This function takes a wide format dataset and converts it to a long format dataset.
}
\keyword{internal}
